src/python/various/requirements_mac.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
comm==0.1.4
contourpy==1.1.1
cycler==0.12.0
dataclasses-json==0.6.1
datasets==2.14.5
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
dnspython==2.4.2
executing==2.0.0
fastjsonschema==2.18.0
filelock==3.12.4
fonttools==4.43.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
google-ai-generativelanguage==0.3.2
google-api-core==2.12.0
google-auth==2.23.2
google-generativeai==0.2.0
googleapis-common-protos==1.60.0
grpcio==1.59.0
grpcio-status==1.59.0
huggingface-hub==0.17.3
hydra-core==1.3.2
idna==3.4
ipykernel==6.25.2
ipython==8.16.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
langchain==0.0.305
langsmith==0.0.41
llvmlite==0.41.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mistune==3.0.2
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
nltk==3.8.1
notebook==7.0.4
notebook_shim==0.2.3
numba==0.58.0
numexpr==2.8.7
numpy==1.25.2
omegaconf==2.3.0
openai==0.28.1
openai-whisper==20231117
overrides==7.4.0
packaging==23.1
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pymongo==4.6.0
pynndescent==0.5.10
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
rsa==4.9
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.3
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
tensorflow-hub==0.14.0
terminado==0.17.1
threadpoolctl==3.2.0
tiktoken==0.5.1
tinycss2==1.2.1
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
transformers==4.33.3
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
umap-learn==0.5.4
uri-template==1.3.0
urllib3==2.0.5
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
widgetsnbextension==4.0.9
xxhash==3.3.0
yarl==1.9.2
